backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic[email]==2.5.0
redis==5.0.1
celery==5.3.4
flower==2.0.1
httpx==0.25.1
jinja2==3.1.2
python-multipart==0.0.6
pillow==10.1.0
boto3==1.33.0
stripe==7.6.0
sentry-sdk==1.38.0
prometheus-client==0.19.0
slowapi==0.1.9